+++ lib/gtm.ts
@@ -1,6 +1,6 @@
 // lib/gtm.ts
 type WindowWithDataLayer = Window & {
-  dataLayer: Record<string, any>[];
+	dataLayer: Record<string, any>[];
 };
 
 declare const window: WindowWithDataLayer;
@@ -8,15 +8,15 @@ declare const window: WindowWithDataLayer;
 export const GTM_ID = process.env.NEXT_PUBLIC_GTM_ID;
 
 export const pageview = (url: string) => {
-  if (typeof window.dataLayer !== 'undefined') {
-    window.dataLayer.push({
-      event: 'pageview',
-      page: url
-    });
-  } else {
-    console.debug({
-      event: 'pageview',
-      page: url
-    });
-  }
+	if (typeof window.dataLayer !== "undefined") {
+		window.dataLayer.push({
+			event: "pageview",
+			page: url,
+		});
+	} else {
+		console.debug({
+			event: "pageview",
+			page: url,
+		});
+	}
 };
